Move onPressReply into child component
